refactor(namespace): apply new  pattern
