+++ infra/test/tsconfig.type-test.json
@@ -2,7 +2,7 @@
     "extends": "../build/tsconfig.base.json",
     "include": [
         "../../**/*.test-d.ts",
-        "../../packages/common/namespace/internal/src/isValidNamespace.ts",
+        "../../**/*.testable-type.ts",
         "../../infra/test/utils/testGroups.ts"
      ],
      "exclude": [
