+++ packages/common/namespace/internal/index.ts
@@ -1,2 +1,3 @@
 export { createNamespacePrepender, type Namespaced } from './src/createNamespacePrepender';
-export { isValidNamespace, type ValidNamespace } from './src/isValidNamespace';
+export { isValidNamespace } from './src/isValidNamespace';
+export { type ValidNamespace } from './src/ValidNamespace.testable-type';
